Fix mypy 1.8 errors: guard audio loaders against load() failure; cast cv2 RGB frames to uint8 for DetectObjectsFrameInput/tracking append

+++ model-service/src/infrastructure/adapters/inbound/fastapi/routes/detection.py
@@ -8,6 +8,7 @@ from __future__ import annotations
 import logging
 
 import cv2
+import numpy as np
 from fastapi import APIRouter, HTTPException
 from opentelemetry import trace
 
@@ -114,7 +115,7 @@ async def detect_objects(
                 ret, frame = cap.read()
                 if not ret:
                     continue
-                frame_rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
+                frame_rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB).astype(np.uint8)
                 timestamp = frame_num / fps if fps > 0 else 0.0
                 frame_inputs.append(
                     DetectObjectsFrameInput(
